+++ Assets/Scripts/Entities/EntityFactory.cs.meta
@@ -1,11 +1,3 @@
 fileFormatVersion: 2
-guid: fe8b2a95fd283eb498bc8e1aecb2a39a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: a43901eb5d77410da2f9f850b67beb18
+timeCreated: 1716723402
