--- Assets/Scripts/Entities/EntityFactoryManager.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a43901eb5d77410da2f9f850b67beb18
-timeCreated: 1716723402
